+++ packages/ionic/src/lib/session.ts
@@ -137,7 +137,7 @@ export async function promptToLogin(env: IonicEnvironment): Promise<void> {
   const { validators } = await import('@ionic/cli-framework');
 
   env.log.msg(
-    `Log into your Ionic Pro account\n` +
+    `Log into your Ionic Appflow account\n` +
     `If you don't have one yet, create yours by running: ${chalk.green(`ionic signup`)}\n`
   );